pyproject.toml
[project]
name = "llm-lsp-cli"
version = "1.0.0"
description = "A CLI tool that interacts with language servers to provide code intelligence features for LLM-assisted development"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Rianico", email = "zhexunakun@163.com" }]
keywords = ["lsp", "language-server", "cli", "code-intelligence", "llm", "developer-tools"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development",
  "Topic :: Text Editors :: Integrated Development Environments (IDE)",
]
dependencies = [
  "typer>=0.9.0",
  "python-daemon>=3.0.0",
  "pydantic>=2.5.0",
  "psutil>=5.9.0",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/Rianico/llm-lsp-cli"
Repository = "https://github.com/Rianico/llm-lsp-cli"
Issues = "https://github.com/Rianico/llm-lsp-cli/issues"

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.1.0",
  "mypy>=1.7.0",
  "types-PyYAML>=6.0",
]

[project.scripts]
llm-lsp-cli = "llm_lsp_cli.cli:app"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_lsp_cli"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest-timeout>=2.4.0",
]
